fix : recipe edit issue solved

+++ src/main/java/com/MealMate/MealMateBackend/recipe/dto/IngredientItem.java
@@ -4,13 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class IngredientItem {
     private String name;
-    private BigDecimal quantity;
+    private Double quantity;
     private String unit;
 }
